+++ api/src/endware/Archive/index.ts
@@ -1,5 +1,7 @@
 import archiveReport from "./archiveReport"
+import unarchiveReport from "./unarchiveReport"
 
 export {
-  archiveReport
+  archiveReport,
+  unarchiveReport
 }