+++ api/src/middleware/index.ts
@@ -3,11 +3,13 @@ import verifyToken from "./verifyToken";
 import verifyVersion from "./verifyVersion";
 import verifyReportFileID from "./verifyReportFileID";
 import verifyReportFileID_min from "./verifyReportFileID_min";
+import verifyArchiveID from "./verifyArchiveID";
 
 export default { 
   verifyToken,
   verifyVersion,
   verifyReportID,
   verifyReportFileID,
-  verifyReportFileID_min
+  verifyReportFileID_min,
+  verifyArchiveID
 };