+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### --- v0.3.8 --- 05.06.23, 12.42 UTC+2 ---
+* API: Added: Archived Reports can now be "unachived"
+* API: Added: Middleware check for archived reports if they exist
+
 ### --- v0.3.7 --- 05.06.23, 12.34 UTC+2 ---
 * API: Added: Reports can now be archived